Add middleware to authenticate user
Currently there is an issue where the access token is not checked
for expiry which means it will respond with the wrong error payload.
The ValidateAccessToken method needs to be updated to properly return
an error when a token is expired. This also requires updating the tests
for the function to properly test for specific errors.