Refactor the authentication middleware to refresh the access token
This commit refactors the auth middleware to attempt a refresh of the access token if the access token or cookie is expired. Any other issues with the access token or its cookie will cause authentication to fail.

Additionally, I have also refactored the authedUser struct to contain an Error field instead of an IsAuthed field. This ensures handlers can handle specific failed authentications if needed.